update(config/org.yaml): bulk sync from GH APIs
Note that in teams, some persons are listed under `maintainers` instead of `members` because they are organization admins. In this context, `maintainers` means "maintainers of the team" (not of the repo). Moving them to `members` is ineffective for GH. Unfortunately, this may be misleading, but there's nothing we can do.

Signed-off-by: Leonardo Grasso <me@leonardograsso.com>